Integrate with Flyway (Closes #31)
This commit integrates with Flyway as the migration tool for the database instead of relying on `Exposed` to create the database schema based on code.

Moreover, the domain and persistence layer have been updated to store the dates as numbers and not strings

+++ fetcher/src/main/kotlin/com/xkcddatahub/fetcher/adapters/output/http/rest/xkcd/mappers/ApiWebComicsMapper.kt
@@ -7,9 +7,9 @@ object ApiWebComicsMapper {
     fun ApiWebComics.toDomain() =
         WebComics(
             id = id,
-            year = year,
-            month = month,
-            day = day,
+            year = year.toInt(),
+            month = month.toInt(),
+            day = day.toInt(),
             title = title,
             safeTitle = safeTitle,
             transcript = transcript,
